requirements.txt
streamlit
langchain
langgraph
langchain-google-vertexai
pydantic
typing-extensions
pandas
google-cloud-bigquery
google-cloud-bigquery[pandas]
google-cloud-bigquery-storage
google-cloud-storage
google-cloud-aiplatform